Refactor Account class and AccountService to use long long for account numbers; add transfer method for inter-account transactions.

+++ include/models/Account.h
@@ -2,14 +2,14 @@
 
 class Account{
     private:
-        long accountNumber;
+        long long accountNumber;
         long long customerId;
         double balance;
 
     public:
-        Account(long accountNumber, long long customerId, double balance): accountNumber(accountNumber), customerId(customerId), balance(balance) {}
+        Account(long long accountNumber, long long customerId, double balance): accountNumber(accountNumber), customerId(customerId), balance(balance) {}
        
-        long getAccountNumber() const;
+        long long getAccountNumber() const;
         long long getCustomerId() const;
         double getBalance() const;
         void deposit(double amount);
